requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.0
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
python-socketio==5.9.0
aiortc==1.5.0
reportlab==4.0.4
img2pdf==0.4.4
google-cloud-secret-manager==2.16.0
pytest==7.4.3
pytest-cov==4.1.0
